requirements.txt
click==7.1.2
Flask==1.1.2
Flask-MySQLdb==0.2.0
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
mysql-connector-python==8.0.20
mysqlclient==1.4.6
protobuf==3.11.3
six==1.14.0
Werkzeug==1.0.1
